use double to be TimeType
git-svn-id: svn://222.29.86.177/wanghan/projects/gpu-md/branches/trials/improved-neighbor-list@465 a28d8dd4-3dfb-4ec0-b98a-e5cf011a4b41

+++ include/MDTimer.h
@@ -3,7 +3,7 @@
 
 #define NumberOfMemberInMdTime 32
 
-typedef float		TimeType;
+typedef double		TimeType;
 
 
 enum mdTimeItem {
